fine-tuning lmssetupContainer - final changes.

+++ scripts/usr/local/lib/lms/lmssetupContainer-0.0.2.bash
@@ -84,11 +84,19 @@ function setupContainer()
 
 	for fname in *
 	do
-    	[[ -f "${fname}" ]] && rmDestFile "/usrlocal/bin/${fname}"
-    	cp -r "${fname}" "/usrlocal/bin/${fname}"
+    	[[ -f "/usrlocal/bin/${fname}" ]] && 
+    	 {
+    	    if ! [ -L "/usrlocal/bin/${fname}" ]
+    	    then 
+    	        rmDestFile "/usrlocal/bin/${fname}"
+    	        cp -r "${fname}" "/usrlocal/bin/${fname}"
+    	    fi
+    	 }
+    	
 	done
 
-	chmod +x /usrlocal/bin/*
+	chmod +x /usrlocal/bin/debian-*
+	chmod +x /usrlocal/bin/docker*
 
 	echo "LMS_BASE=${LMS_BASE}" > /conf/lms-base.conf
 	chmod 755 /conf/lms-base.conf
